float para int

+++ app.py
@@ -51,7 +51,7 @@ def predict():
         result = model_et.predict(data_df)
         print("Result:", result)
         # send back to browser
-        output = {'results': float(result[0])}
+        output = {'results': int(result[0])}
         #output = {'results': float(status)}
 
     # return data
